requirements.txt
flask
flask-cors
sqlalchemy
waitress
requests
psycopg2-binary
python-dotenv
